STYLE: Remove deprecated cmake directives.
The online help says that we should use ADD_SUBDIRECTORY
rather than SUBDIRS.

<quote  author="Bill Hoffman">
SUBDIRS is processed at the end of the current CMakeLists.txt file.
ADD_SUBDIRECTORY is done immediately. So, if you set
variables in the parent directory they will not be seen in the subdir
when
SUBDIRS is used.
</quote>